Removed animation
removed animation and looking at changing event handling. Character
animation is handle by a timer

+++ SDL/Time.h
@@ -21,9 +21,9 @@ public:
 	bool isStarted();
 
 private:
-	Uint32 iStartTick;
-	Uint32 iPausedTick;
-	Uint32 iLastTick;
+	float iStartTick;
+	float iPausedTick;
+	float iLastTick;
 
 	float fDelta;
 	bool bPaused;
